# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loreline",
    "summary": "Tabletop session transcriber - capture + STT orchestration.",
    "version": "1"
  },
  "paths": {
...
    "/api/campaigns": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "List Campaigns",
        "description": "Every campaign, by name, with its session count and last session.",
        "operationId": "list_campaigns_api_campaigns_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignSummary"
                  },
                  "type": "array",
                  "title": "Response List Campaigns Api Campaigns Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Create Campaign",
        "description": "Create a campaign.\n\nA name already in use is a 409 rather than a second row: the picker shows\nnames, so two campaigns called \"Curse of Strahd\" are indistinguishable\neverywhere a GM would have to choose between them, and the mistake that\nproduces them is pressing New twice.",
        "operationId": "create_campaign_api_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get Campaign",
        "description": "One campaign.",
        "operationId": "get_campaign_api_campaigns__campaign_id__get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "campaigns"
        ],
        "summary": "Update Campaign",
        "description": "Rename a campaign, or change its notes or its recap prompt.",
        "operationId": "update_campaign_api_campaigns__campaign_id__put",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "campaigns"
        ],
        "summary": "Delete Campaign",
        "description": "Delete a campaign. Its sessions are kept, unassigned.\n\nSaid in the confirm dialog too, because it is the one thing about this\nbutton anybody could be afraid of: a campaign is a label on recordings that\nalready exist, and deleting the label must not be able to delete four hours\nof audio nobody can capture again.",
        "operationId": "delete_campaign_api_campaigns__campaign_id__delete",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/sessions": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Campaign Sessions",
        "description": "The campaign's sessions, oldest first - the order they were played in.\n\nThe opposite of the History page's order, deliberately: that page answers\n\"what did I record last\", and a campaign answers \"what happened, and then\nwhat happened\", which is a story and reads forwards.",
        "operationId": "campaign_sessions_api_campaigns__campaign_id__sessions_get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  },
                  "title": "Response Campaign Sessions Api Campaigns  Campaign Id  Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/documents": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Campaign Session Documents",
        "description": "Every generated text of the campaign's sessions, oldest session first.",
        "operationId": "campaign_session_documents_api_campaigns__campaign_id__documents_get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionDocument"
                  },
                  "title": "Response Campaign Session Documents Api Campaigns  Campaign Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/entities": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Campaign Entities",
        "description": "Every session's extraction, merged by name.\n\nMerged on a normalised name (trimmed and case-folded) rather than on the\nexact string, because a transcript spells a name three ways across nine\nsessions and three entries for one NPC is exactly the mess this page exists\nto replace. The spelling kept is the first one seen, the notes are the\nlatest session's, and the session list says where the name came up - which\nis what turns \"who was Vallaki's burgomaster again\" into a link.",
        "operationId": "campaign_entities_api_campaigns__campaign_id__entities_get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignEntities"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/glossary/add": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Add To Campaign Glossary",
        "description": "Append names to the campaign's glossary, trimmed and deduplicated.\n\nAppend rather than replace: this is reached from the extracted-names list,\nwhere the gesture is \"and these too\", and a request that could silently\ndrop the terms a GM typed by hand would make that list dangerous to use.\nCase-insensitive on the duplicate check, because \"Strahd\" and \"strahd\" bias\na recognizer identically and two of them only spend the glossary ceiling.",
        "operationId": "add_to_campaign_glossary_api_campaigns__campaign_id__glossary_add_post",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlossaryAdd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Glossary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaign_id}/previously-on": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get Previously On",
        "description": "The stored \"previously on\", or null when none has been written.",
        "operationId": "get_previously_on_api_campaigns__campaign_id__previously_on_get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CampaignDocument"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Previously On Api Campaigns  Campaign Id  Previously On Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Write Previously On",
        "description": "Write the campaign's \"previously on\" from its last few sessions.\n\nReads recaps, falling back per session to the stored summary, because a\ncampaign is rarely all one or all the other: the last three sessions have\nrecaps and the one before them was summarized months ago, and refusing to\nread that one would silently drop a session out of the story. A session\nwith neither is skipped rather than represented by its raw transcript -\nfour hours of dialogue would swamp the two paragraphs either side of it.",
        "operationId": "write_previously_on_api_campaigns__campaign_id__previously_on_post",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviouslyOnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/session/{session_id}/campaign": {
      "put": {
        "tags": [
          "sessions"
        ],
        "summary": "Set Session Campaign",
        "description": "Put a session in a campaign, or take it out of one (null).\n\nAn id no campaign answers to is a 404 rather than a stored string: that is\nexactly the state this feature exists to end - a ``campaign_id`` nothing\ncan resolve, rendering as a raw hex id in the History table and reaching a\nglossary nobody can find.",
        "operationId": "set_session_campaign_api_session__session_id__campaign_put",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignAssignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignSummary": {
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/Campaign"
          },
          "sessions": {
            "type": "integer",
            "title": "Sessions",
            "default": 0
          },
          "last_session_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session At"
          }
        },
        "type": "object",
        "required": [
          "campaign"
        ],
        "title": "CampaignSummary",
        "description": "A campaign as the list page needs it: the row, plus what it holds.\n\nThe two counts are the whole reason the list is not just the rows: \"which\nof these am I actually playing\" is answered by how many sessions it has and\nwhen the last one was, and neither is on the campaign itself."
      },
      "CampaignWrite": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "recap_prompt": {
            "type": "string",
            "title": "Recap Prompt",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CampaignWrite",
        "description": "Create or rename a campaign.\n\nThe name is the campaign, so a blank one is refused rather than stored: a\nrow that renders as an empty cell in the picker is unpickable and\nindistinguishable from \"no campaign\", which is the one thing the list must\nbe able to say."
      },
...
      "GlossaryAdd": {
        "properties": {
          "terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Terms"
          }
        },
        "type": "object",
        "title": "GlossaryAdd",
        "description": "Names to append to a campaign's glossary, deduplicated on the way in."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "OkResponse",
        "description": "Generic success acknowledgement."
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session JWT set by `POST /api/auth/login`. The cookie is HttpOnly, so pasting a value into this dialog does nothing: browsers drop the `Cookie` header Swagger UI would set from it. To use `Try it out`, run `POST /api/auth/login` from this page first (or log into the app in the same browser); the cookie is then sent automatically. Outside a browser, send the cookie yourself, e.g. `curl -c jar -X POST .../api/auth/login -d '{\"password\":\"...\"}' -H 'content-type: application/json'` then `curl -b jar .../api/session`. When no password is configured auth is off entirely and the cookie is not required.",
        "in": "cookie",
        "name": "loreline_token"
      }
    }
  }
}